Implemented pedantic static analysis on Windows.
- Enabled all VC++ warnings (including level 4 warnings and
  the ones disabled by default), selectively disabled noisy
  and unfeasible ones.
- Warnings are now treated as errors.
- Disabled the support for exception handling.
- Bumped up the level of basic security checks, enabled SDL.
- Isolated C++/CX to POSIX emulation.
- Enabled PREfast.
- Turned on analysis in CSSOM and fixed detected issues.

+++ cobalt/math/transform_2d_test.cc
@@ -52,7 +52,7 @@ TEST(Transform2DTest, IsotropicScale) {
 }
 
 TEST(Transform2DTest, Rotation) {
-  const float kTheta = M_PI / 4.0f;
+  const float kTheta = static_cast<float>(M_PI) / 4.0f;
 
   // Note that we account for the fact that up is represented by negative y.
   Matrix3F expected_value = Matrix3F::FromValues(
